采用万得数据导出的成分股权重计算上证50和沪深300指数，并将期货由IF修改为IH，将显示界面做了调整。

+++ HiStar/me.h
@@ -9,5 +9,5 @@ struct sqldb{
 struct stock{
 	std::string exch;
 	std::string code;
-	int volume;
+	double volume;
 };
